fix: exception frame changed feat: zUART now prints via IPC (reg-reg only)
zUART v1.0.0 is now fully able to write to UART
Exception frame saves sp and lr altogether

+++ core/log.h
@@ -8,11 +8,11 @@
 #define LOG_FMT(fmt) fmt
 #endif
 
-#define KINFO(fmt, ...)  kprintf("\033[36m[%6llu INFO ] \033[0m" LOG_FMT(fmt) "\n", get_ticks(), ##__VA_ARGS__)
-#define KWARN(fmt, ...)  kprintf("\033[33m[%6llu WARN ] \033[0m" LOG_FMT(fmt) "\n", get_ticks(), ##__VA_ARGS__)
-#define KERROR(fmt, ...) kprintf("\033[31m[%6llu ERR  ] \033[0m" LOG_FMT(fmt) "\n", get_ticks(), ##__VA_ARGS__)
+#define KINFO(fmt, ...)  kprintf("\033[36m[%6llu INFO ] \033[0m" LOG_FMT(fmt) "\n", (unsigned long long)get_ticks(), ##__VA_ARGS__)
+#define KWARN(fmt, ...)  kprintf("\033[33m[%6llu WARN ] \033[0m" LOG_FMT(fmt) "\n", (unsigned long long)get_ticks(), ##__VA_ARGS__)
+#define KERROR(fmt, ...) kprintf("\033[31m[%6llu ERR  ] \033[0m" LOG_FMT(fmt) "\n", (unsigned long long)get_ticks(), ##__VA_ARGS__)
 #ifdef DEBUG_PRINT
-#define KDEBUG(fmt, ...) kprintf("\033[35m[%6llu DEBUG] \033[0m" LOG_FMT(fmt) "\n", get_ticks(), ##__VA_ARGS__)
+#define KDEBUG(fmt, ...) kprintf("\033[35m[%6llu DEBUG] \033[0m" LOG_FMT(fmt) "\n", (unsigned long long)get_ticks(), ##__VA_ARGS__)
 #else
 #define KDEBUG(fmt, ...) do {} while(0)
 #endif
